% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_study_state.R
\name{boxplot_study_state}
\alias{boxplot_study_state}
\title{Provide a boxplot of study data}
\usage{
boxplot_study_state(data)
}
\arguments{
\item{data}{data from education 2014}
}
\value{
a ggplot object
}
\description{
Provide a boxplot of study data
}
\examples{
# no example data yet
}
