% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_age_groups.R
\name{clean_age_groups}
\alias{clean_age_groups}
\title{Clean age group data}
\usage{
clean_age_groups(age_groups)
}
\arguments{
\item{age_groups}{age group data with \code{-} or \verb{--} or \verb{---} separating the
numbers.}
}
\value{
numbers separated by \verb{_}
}
\description{
Turn text from \code{10--12} or \code{10-12} into \verb{10_12}
}
\examples{
clean_age_groups(raw_education_2014)
}
