% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_education_data.R
\name{clean_education_data}
\alias{clean_education_data}
\title{Cleaning 2014 education data}
\usage{
clean_education_data(data)
}
\arguments{
\item{data}{data from education 2014}
}
\value{
cleaned up data, with NAs identified, and age groups cleaned
}
\description{
Cleaning 2014 education data
}
\examples{
# not used yet
}
