% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learned-package.R
\docType{package}
\name{learned-package}
\alias{learned}
\alias{learned-package}
\title{learned: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

}
\keyword{internal}
