% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_study_age_state.R
\name{plot_study_age_state}
\alias{plot_study_age_state}
\title{Plot study against age and state}
\usage{
plot_study_age_state(data)
}
\arguments{
\item{data}{education 2014 data}
}
\value{
ggplot
}
\description{
Plot study against age and state
}
\examples{
plot_study_age_state(raw_education_2014)
}
