% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_prop_study.R
\name{summarise_prop_study}
\alias{summarise_prop_study}
\title{Provide a summary of the proportion studying}
\usage{
summarise_prop_study(data)
}
\arguments{
\item{data}{education 2014 data}
}
\value{
5 number summary containing min, q1, median, q3, and standard deviation
}
\description{
Provide a summary of the proportion studying
}
\examples{
summarise_prop_study(raw_education_2014)
}
